test(guest-download): add integration tests with mock http server
split main.rs into lib.rs + main.rs to expose run() for testing.
add 14 integration tests covering download, extract, retry, parallel,
and error handling paths using httpmock + tempfile.
